Improved default audio/subtitle priority
The problem is that sometimes "eng" is used to mean English and other times "en" is used, so here we use both when they use the English UMS interface.
It's not a comprehensive fix; a better fix would be to make all occurrences of "en" also apply to "eng", so maybe we can do that later. This should work in the meantime.