Removed all default values from UMS.conf
The program uses its internal defaults when it can't find a value in UMS.conf.
When a user saves settings, usually the only ones that are saved are the ones that were changed, so that means that with this commit users will now be able to take advantage of our new defaults when we change them unless they have changed them themselves.
Prior to this commit, the user would never get updated settings unless they deleted UMS.conf or changed them manually.