FFmpeg can pass the video to tsMuxeR (via MEncoder) if compatible, like MEncoder already does
We should probably continue this by making tsMuxeRVideo able to use FFmpeg to pass it the streams, to allow the user to completely bypass MEncoder if they want. That will depend on how stable tsMuxeR is with FFmpeg output though, since we already know tsMuxeR and MEncoder get along extremely well.