New TranscodeVideo option: H264TSAAC
I created this option primarily to make UMS compatible with my Sony
Xperia C6603. In my tests, this option performed very well, although AAC
is not completely supported, by FFmpeg, MEncoder (no support at all in
the version distributed with UMS) and VLC. But it was good enough to me.