Fix FFmpeg subtitle convert
Added a stopProcess() at the end of FFmpeg subtitle SRT to ASS
converting process. Without this, a pipe is created and a timer keeps
messing up the buffer progress bar, showing 0 MB/XX MB intermittently.
We can see this happening both in GUI interface and trace log when
playing a video with on-the-fly conversion of SRT subtitle to ASS with
FFmpeg.