Updated MediaInfo for Windows to 0.7.71
I conducted a speed test on a folder with 100+ videos and found that with our current settings, 0.7.71 finished scanning an average of 5% faster than 0.7.69.
It may have just been chance (I only did 4 tests for each version), or the result of hard disk caching, but I think we can at least say with certainty that the newer one isn't slower.