Migrates from Netty 3.x to 4.0.25.Final
In Netty 3.x, if the kernel's send buffer fills up, it will suspend
writing data to the client until the stream is finished being generated.
In the case of transcoding large videos, that may be (and often is)
longer than the TV is willing to wait before closing its request and
opening a new one. The issue is not present in Netty 4.x.

Note, this is a naive migration to Netty 4.x. It should probably go
through some more testing and incubation before being released.

Closes #454