Isolates getting the SOAPAction header
This pulls the logic for parsing out the SOAPAction value into its own
method. In the previous version, it would be set to the last header
value of a SOAPAction or Callback header. Isolating the logic into its
own method should make it more clear what the behavior is and easier to
change if it needs to be updated in the future.

It should be okay to assume that the header value is non-empty if the
header is set because had that not been the case, the previous code
would have thrown a NoSuchElementException when calling .nextToken().