DLNA standard
cf. upnp-av-contentdirectory-v3-service.pdf
For UPnP Device Architecture defined boolean data types, it is strongly RECOMMENDED to use the 
value “0” for false, and the value “1” for true. However, when used as input arguments, the values “false”, 
“no”, “true”, “yes” may also be encountered and MUST be accepted. Nevertheless, it is strongly 
RECOMMENDED that all boolean state variables and output arguments be represented as “0” and “1”. 
For XML Schema defined Boolean data types, it is strongly RECOMMENDED to use the value “0” for 
false, and the value “1” for true. However, when used as input properties, the values “false”, “true” may 
also be encountered and MUST be accepted. Nevertheless, it is strongly RECOMMENDED that all 
Boolean properties be represented as “0” and “1”.