Better structured filename re-writing testcase (#1630)
* Better structured filename re-writing testcase

This patch factors the expected behaviour of the filename prettifier out
of the testcase code and puts it into a JSON resource. The testcase
code goes over the test definitions in the JSON and executes them.

* Update pom.xml

Co-Authored-By: stasinos <stasinos@konstant.gr>

* Update src/test/java/net/pms/util/FileUtilTest.java

Co-Authored-By: stasinos <stasinos@konstant.gr>

* Fix indentation error

* Minor additions to filename prettifier testcases

Added two new testcases that check behaviour when movie filenames
have a year as part of the title.
Exceptions that indicate a brolen installation cause are encapsulated
in an assertion error, causing the test to fail.