A possibility to have a virtual folder without adding it to the Media Library.
Fixed typo. Very sorry.

Documentation of the feature. Please note that the feature will only
work if you use the configuration in a separate file
(VirtualFolders.conf). The "virtual_folders" variable directly in
UMS.conf will always include folders in the Media Library. If not, the
syntax would be a real mess.