Fixed an undesirable behavior inserted with initial commit of extended MRR compatibility, such that the phrase 'Windows Media Connect' was being appended to the server name.
This was caused by the new default behavior which included simulation of MRR service to renderers. This in turn activated the legacy behavior for Xbox 360 for which the UMS server name was modified.

This commit reverts to the -legacy- behavior such that the server name is only decorated for the Xbox 360.

Note that special service to the Xbox 360 is activated by the presence of the 'xbox 360' in the renderer name. Any renderer which includes this phrase in its name via the config option, 'RendererName', will obtain the same 'special' service.