Fixed failed database writes
There were some cases of a method establishing a shared db connection, then calling another method which altered that connection using setAutoCommit, which caused unexpected results.
In this commit I have tried to establish a pattern where the method that establishes the Connection, is also the one that alters it, which gives better visibility over the state of the connection.
Along the way I did minor optimizations that shave milliseconds.