Stop treating LG video thumbnail requests as play/stop events
LG TVs send play requests in their Media Player app to get dynamic thumbnails on hover, which causes all kinds of problems, like marking the video as fully played if the thumbnail requested is near the end of the video. Here we detect the unique user-agent of those requests and treat them differently.