Update FFMpegVideo.java
A user on the forum https://www.universalmediaserver.com/forum/viewtopic.php?t=15646 was asking about the error "Couldn't parse any supported protocols for" in the logs. I get that error too.

They think it is due to this line of code and to me it looks like the logic is inverted if I'm reading it right.

This is the first time I tried this so don't know if I have done it right.