Update SubtitleUtils.java: additional flags on external subtitle files
Add the following markers:
default: Marks the stream as the default.
forced, foreign: Marks the subtitle stream as forced, typically used for translation of segments of audio/text that differ from the primary language.
sdh, cc: Indicates that the subtitle stream has additional information to help viewers that are hearing impaired. This is at the moment without integration as it seems there is no possibility to mark a subtitle for that (maybe in the future?) Note that it doesn't support the "hi" keyword as this clashes with the "hi" marker for the hindi language.
It seems it is a common standard to mark external subtitle files that way (in Jellyfin but also on other media servers). See: https://jellyfin.org/docs/general/server/media/external-files/